move worldDataObject to data folder
